Patient-Triage-RAG/requirements.txt
chromadb
fastapi
groq
httpx
huggingface-hub
python-dotenv
python-multipart
requests
rich
sentence-transformers
uvicorn
slowapi